+++ pkg/driver/hook/hook.go
@@ -1,4 +1,4 @@
-package hooks
+package hook
 
 import (
 	"context"
