+++ pkg/driver/hook/options.go
@@ -1,4 +1,4 @@
-package hooks
+package hook
 
 import (
 	"os"
