+++ pkg/driver/hook/prestop.go
@@ -1,4 +1,4 @@
-package hooks
+package hook
 
 import (
 	"context"
